fix: CI cross-arch build failures
- x86 raw syscall: syscall number i64 does not match the 32-bit eax
  register class — truncate to u32
- variant stub build passed the target triple without a --target flag
  in front of it
- rustup triples for the Android targets carry no API level: use
  armv7-linux-androideabi / aarch64-linux-android

+++ crates/shell_stub/src/raw.rs
@@ -154,7 +154,7 @@ pub unsafe fn syscall5(n: i64, a: usize, b: usize, c: usize, d: usize, e: usize)
     let ret;
     core::arch::asm!(
         "int $$0x80",
-        in("eax") n,
+        in("eax") n as u32,
         in("ebx") a, in("ecx") b, in("edx") c, in("esi") d, in("edi") e,
         lateout("eax") ret,
         options(nostack)
